requirements.txt
# Core dependencies
numpy>=1.21.0
opencv-python>=4.5.3
Pillow>=8.3.1
pytesseract>=0.3.8

# GUI
tk>=0.1.0

# Data handling
pyyaml>=6.0
jsonschema>=4.0.0

# Utilities
tqdm>=4.62.0
python-dotenv>=0.19.0

# LLM Integration (optional)
# openai>=0.27.0
# anthropic>=0.3.0